pyproject.toml
[tool.poetry]
name = "it-assets"
version = "2.3.58"
description = "DBCA IT assets (both physical and knowledge-based) management system"
authors = ["DBCA OIM <asi@dbca.wa.gov.au>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
psycopg2-binary = "2.9.1"
dbca-utils = "1.1.6"
django-extensions = "3.1.3"
django-dotenv = "1.4.2"
dj-database-url = "0.5.0"
gunicorn = "20.1.0"
openpyxl = "3.0.5"
unicodecsv = "0.14.1"
python-dateutil = "2.8.1"
django-leaflet = "0.28.2"
Pillow = "9.0.1"
restless = "2.1.1"
XlsxWriter = "1.3.3"
django-crispy-forms = "1.10.0"
texttable = "1.6.3"
titlecase = "0.12.0"
data-storage = {git = "https://github.com/dbca-wa/data-storage.git", rev = "v3.18"}
django-bootstrap-pagination = "1.7.1"
django-markdownx = "3.0.1"
webtemplate-dbca = "1.5.0"
fuzzywuzzy = "0.18.0"
python-Levenshtein = "0.12.0"
mixer = "6.1.3"
msal = "1.12.0"
django-json-widget = "^1.1.1"
PyYAML = "5.4.1"
whitenoise = {version = "5.3.0", extras = ["brotli"]}
pysftp = "0.2.9"
django-csp = "3.7"
django-sockpuppet = {extras = ["lxml"], version = "^0.6.0"}

[tool.poetry.dev-dependencies]
ipython = "^7.31.1"
ipdb = "^0.13.3"

[build-system]
requires = ["poetry>=1.1.13"]
build-backend = "poetry.masonry.api"
